refactor(tools): split monolithic tools.ts into modular tools/ directory
Separate 2105-line tools.ts into bridge.ts (shared infra), shared.ts
(bot + userbot tools), and userbot.ts (MTProto-only tools). The new
tools/index.ts conditionally registers userbot tools based on
TALON_FRONTEND_MODE env var passed from the SDK launcher.

Remove 3 deprecated tools: edit_last_message, get_premium_gift_options,
check_privacy_for_user (and their corresponding action handlers).
